requirements.txt
streamlit
python-docx
python-pptx
dashscope